SplitModel initial work
In order to work on multi-models, you have to split up a model
into consituent parts.

This is so that the function that updates the states for a particular
time period can be seperated.

I had originally thought there needed to be 3 different sections of a
model (setup, update and report), I realised the classic model for the
lake runs a few variants of the model, and so needs something to setup a
variant (variant_setup)  and also modify the report based on the outcome of that variant (variant_report).

I've attempted to re-implement the classic lake model in the new split
variant. It is a little messier, and harder to understand.

Things that can be improved:

- Use reflection to look the variables required for a function and only
pass them in from the state. This would make them a lot cleaner. If
a plain value needs to be mutated, return a dictionary with it in and
merge it with the state?

- If there isn't variant setup/report phases don't run them.

- Tests to make sure that this new lake model says the same as the old
  ones.