update IntegerParameter
Suggestion to also accept float values that are verified integers. long() and int() are unified in python 3 and therefore the use of the integer instance shouldn't be a problem in comparison to the numbers.Integral instance. This suggestion makes it possible to use real and int parameters in the create_parameters function when loading a csv-file. This was/is causing issues as integers are transformed to floats when the bound values of these parameters where mixed.