update utilities.experiments_to_scenarios to datafames
Hi Jan, 

When trying to rerun experiments using experiments_to_scenarios I found that the format that is expected is not consistent with the format used by ema_workbench.load_results. It expects a structured numpy array while the experiments are returned by load_results as dataframe. I've changed the description, the way of extracting the uncertainty names (now based on simple df.columns) and added type-hinting for the experiments. 

I've tested it in my own setup and it works fine. Let me know if this is okay to change. 

Seth