feature_scoring.py: Change max_features in get_rf_feature_scores to "sqrt"
The usage of max_features="auto" as parameter value in RandomForestRegressors and ExtraTreesRegressors is depreciated in scikit-learn 1.1 and will be removed in 1.3. The default value for max_features also changed from from "auto" to "sqrt" in version 1.1.

This commit changes max_features parameter to "sqrt". Behaviour should be identical. Note that this value can also be removed, since the default value is also "sqrt".