Allow MultiprocessingEvaluator to initialize with cpu_count minus N
This commit allows inputting negative integers for n_processes in the MultiprocessingEvaluator defined in evaluators.py. This way the number of logical CPU cores minus that negative number would be used. For example, on a 12 thread processor, -2 results in using 10 threads. There is a guard around the function that always at least one process.

The main advantage is that when inputting a negative number, there is always some CPU power available for system and background tasks, which means the system would lag way less. When working in teams or on different devices, you can just leave it to -1 or -2 and always have some computing power to spare.