Migrate to mpi4py.futures, simplify bash script
Migrate the Python script from mpi4py.MPI to mpi4py.futures. Note that mpi4py.futures requires a separate launcher script to create the MPI environment and launch the main script.

When using mpi4py.futures, the MPIPoolExecutor needs to be started in a separate process. This is because the MPIPoolExecutor must be created and submitted to tasks within an if __name__ == '__main__': block to prevent recursive creation of MPIPoolExecutor instances.

Also, in the bash commands:
- Modify the bash code to remove or clear the python-test directory, if it already exists. This ensures a clean environment for every test run.
- Simplify the scp file transfer code, to one line.