evaluators.py: Correctly do conditional import
If you only need it in one method of one class, just import it in that one method of that one class (and not in the __init__).

Makes sure mpi4py is only required when the MPIEvaluator is used. Challenging part will be conditionally importing MPI to establish the rank based on which evaluator is used.