em_framework: Improve log messages, warning and errors (#300)
Improve the log messages, warning and errors, by:
- Making them more explicit
- Including what the variable value or type not expected currently is
- Including what the variable value or type not expected should be
- Stating the explicit type of warning or error (DeprecationWarning, ValueError, etc.)
- Converting them to F-strings (for better in-code readability)
- Formatting warnings and errors with a Capital letter, while leaving log-statements lowercase.

This will make debugging easier, allowing more time for model development and experimentation.

There are two placed where it's mentioned that a function is deprecated, but not when they will be removed. For both it's decided that they will be removed in version 3.0, which is added to the log message.