Fixes for various warning during tests (#383)
* fix for pandas futurewarning

fixes `FutureWarning: The default value of observed=False is deprecated and will change to observed=True in a future version of pandas. Specify observed=False to silence this warning and retain the current behavior`

* fix for pandas futures warning

FutureWarning: ChainedAssignmentError: behaviour will change in pandas 3.0!

* fixes platypus deprecation warning for default_variator

* fix for platypus deprecation warning

AutoAdaptiveTimeContinuation is renamed to AutoAdaptiveTimeContinuationExtension.

* fixes deprecation warning in platypus

EpsilonProgressContinuation is renamed to EpsilonProgressContinuationExtension and moved inside platypus.

* fixes pandas future warning

FutureWarning: ChainedAssignmentError: behaviour will change in pandas 3.0!

* fixes division by zero error

* removes ConvergenceWarnings

* Update test_logistic_regression.py

* Update test_logistic_regression.py

* Update test_logistic_regression.py

* Update test_logistic_regression.py

* Update test_logistic_regression.py